+++ pkg/jobstore/boltdb/store_test.go
@@ -651,7 +651,7 @@ func (s *BoltJobstoreTestSuite) TestSearchJobs() {
 		})
 		require.NoError(t, err)
 		require.Equal(t, 5, len(response.Jobs))
-		require.Equal(t, uint32(1), response.Offset)
+		require.Equal(t, uint64(1), response.Offset)
 	})
 
 	s.T().Run("everything limit", func(t *testing.T) {
